Skeleton w/ArgoCD for cluster-as-a-service component
This component has a dependency on ArgoCD much like the gitops service
does. v1.9 of the openshift-gitops operator isn't sufficient, however,
because it provides ArgoCD v2.7 when v2.10 is required for PostDelete
hooks. This conflict is resolved by excluding the gitops component when
EaaS is enabled since the former is expected to be discontinued soon.

Signed-off-by: Alex Misstear <amisstea@redhat.com>